# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ma premi\u00e8re API",
        "description": "Une API permettant l'acc\u00e8s aux donn\u00e9es de la base via une approche RESTful",
        "termsOfService": "",
        "contact": {
            "name": "Guiks",
            "email": "guillaume.delapre@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/historique_membre": {
            "get": {
                "tags": [
                    "historique"
                ],
                "summary": "Get a member list",
                "description": "Use this method to return a listing of historics.",
                "operationId": "listHistoric",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoriqueMembre"
                            }
                        }
                    },
                    "204": {
                        "description": "The request didn't return any content."
                    }
                }
            },
            "post": {
                "tags": [
                    "historique"
                ],
                "summary": "Create an historic",
                "description": "Use this method to create an historic",
                "operationId": "createHistoric",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_membre",
                        "in": "formData",
                        "description": "Member ID",
                        "required": true,
                        "type": "integer",
                        "maximum": "11"
                    },
                    {
                        "name": "id_seance",
                        "in": "formData",
                        "description": "Seance ID",
                        "required": true,
                        "type": "integer",
                        "maximum": "11"
                    },
                    {
                        "name": "date",
                        "in": "formData",
                        "description": "Seance date",
                        "required": true,
                        "type": "string",
                        "format": "datetime"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Historic created"
                    },
                    "422": {
                        "description": "Fields missing or incorrect."
                    }
                }
            }
        },
        "/historique_membre/{id_historique}": {
            "get": {
                "tags": [
                    "historique"
                ],
                "summary": "Find historic by ID",
                "description": "Returns a single historic",
                "operationId": "showHistoric",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_historique",
                        "in": "path",
                        "description": "ID of historic to return",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Historic not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "historique"
                ],
                "summary": "Delete an historic",
                "description": "Use this method to delete an historicbased on its id.",
                "operationId": "deleteHistoric",
                "parameters": [
                    {
                        "name": "id_historique",
                        "in": "path",
                        "description": "Member ID to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historic deleted"
                    },
                    "404": {
                        "description": "Invalid historic value"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "HistoriqueMembre": {
            "required": [
                "id_membre",
                "id_seance",
                "date"
            ],
            "properties": {
                "id_historique": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_membre": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_seance": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "date": {
                    "type": "date",
                    "format": "date",
                    "default": "2015-08-17"
                }
            },
            "xml": {
                "name": "HistoriqueMembre"
            }
        },
...
    }
}